[WEB-819] Move field meta generation and validation to a shared util
